Final Commit

+++ run.py
@@ -6,24 +6,35 @@ from pymongo import MongoClient
 client = MongoClient('mongodb://localhost:27017/')
 db = client['Biblioteca']
 
-#DONE
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Autor
 @app.route('/insert_author', methods=['POST'])
 def insert_author():
     data = request.get_json()
     nombre = data['name']
     autores = db['Autor']
+
+    if nombre == '':
+        return make_response(jsonify({'success': False}), 500)  # 401 is for Unauthorized
+
     if autores.find_one({"name": nombre}):
         print('El autor ya existe')
+        response = make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
     else:
         autores.insert_one({"name": nombre, "libros": []})
-    return jsonify({'success': True})
+        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
+    return response
 
 @app.route('/update_author', methods=['POST'])
 def update_author():
     data = request.get_json()
     oldName = data['oldName']
     newName = data['newName']
+
+    
+    if oldName == '' or newName == '':
+        return make_response(jsonify({'success': False}), 500)
+
     autores = db['Autor']
     libros = db['Libro']
 
@@ -34,12 +45,20 @@ def update_author():
             libros.update_one({"_id": libro["_id"]}, {"$pull": {"autores": oldName}})
             libros.update_one({"_id": libro["_id"]}, {"$push": {"autores": newName}})
         autores.update_one({"name": oldName}, {"$set": {"name": newName}})
-    return jsonify({'success': True})
+        response = make_response(jsonify({'success': True}), 200)
+    else:
+        response = make_response(jsonify({'success': False}), 400)
+    return response
 
 @app.route('/delete_author', methods=['POST'])
 def delete_author():
+    print("a")
     data = request.get_json()
     nombre = data['name']
+
+    if nombre == '':
+        return make_response(jsonify({'success': False}), 500)  # 401 is for Unauthorized
+
     autores = db['Autor']
     libros = db['Libro']
 
@@ -50,13 +69,13 @@ def delete_author():
             libros.update_one({"_id": libro["_id"]}, {"$pull": {"autores": nombre}})
             libros.update_one({"_id": libro["_id"]}, {"$push": {"autores": "Autor desconocido"}})
         autores.delete_one({"name": nombre})
-   
-    #Aquí se borraría el autor de la base de datos
-    print('Se va a borrar el autor con el nombre: ' + nombre)
-    return jsonify({'success': True})
+        response = make_response(jsonify({'success': True}), 200)
+    else:
+        response = make_response(jsonify({'success': False}), 400)
+    return response
 # endregion
 
-#DONE
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Libro
 @app.route('/insert_book', methods=['POST'])
 def insert_book():
@@ -66,30 +85,42 @@ def insert_book():
 
     nombre = data['name']
     author = data['author']
+    code = data['code']
+
+    if nombre == '' or author == '' or code == '':
+        return make_response(jsonify({'success': False}), 500)
 
     #Modificar los libros en la tabla autores
-    if len(list(autores.find({"name": author}))) > 0:
+    if autores.find({"name": author}):
         autores.find_one_and_update({"name": author}, {"$push": {"libros": nombre}})
     else:
         autores.insert_one({"name": author, "libros": []})
         autores.find_one_and_update({"name": author}, {"$push": {"libros": nombre}})
 
     #Modificar en la tabla
-    if len(list(libros.find({"name": nombre}))) > 0:
+    if libros.find_one({"name": nombre}):
         print('El Libro ya existe')
-        libros.find_one_and_update({"name": nombre}, {"$push": {"autores": author}, "ISBN": data['code']})
+        if author in libros.find_one({"name": nombre})['autores']:
+            return make_response(jsonify({'success': False}), 401)
+        else:
+            if autores.find_one({"name": author}):
+                libros.find_one_and_update({"name": nombre}, {"$push": {"autores": author}})
+                return make_response(jsonify({'success': True}), 200)
+            else:
+                return make_response(jsonify({'success': False}), 400)
     else:
-        libros.insert_one({"name": nombre, "autores": [author], "ISBN": data['code']})
-
-
-    #print('El nombre del libro ingresado es: ' + nombre + 'y lo escribio: ' + author)
-    return jsonify({'success': True})
+        libros.insert_one({"name": nombre, "autores": [author], "ISBN": code})
+        return make_response(jsonify({'success': True}), 200)
 
 @app.route('/update_book', methods=['POST'])
 def update_book():
     data = request.get_json()
     oldName = data['oldName']
     newName = data['newName'] 
+
+    if oldName == '' or newName == '':
+        return make_response(jsonify({'success': False}), 500)
+    
     autores = db['Autor']
     libros = db['Libro']
 
@@ -100,13 +131,17 @@ def update_book():
             autores.update_one({"_id": autor["_id"]}, {"$pull": {"libros": oldName}})
             autores.update_one({"_id": autor["_id"]}, {"$push": {"libros": newName}})
         libros.update_one({"name": oldName}, {"$set": {"name": newName}})
-
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
+    else:
+        return make_response(jsonify({'success': False}), 400)
 
 @app.route('/delete_book', methods=['POST'])
 def delete_book():
     data = request.get_json()
     nombre = data['name']
+
+    if nombre == '':
+        return make_response(jsonify({'success': False}), 500)
 
     autores = db['Autor']
     libros = db['Libro']
@@ -117,13 +152,12 @@ def delete_book():
         for autor in autores_libro:
             autores.update_one({"_id": autor["_id"]}, {"$pull": {"libros": nombre}})
         libros.delete_one({"name": nombre})
-
-    #Aquí se borraría el autor de la base de datos
-    print('Se va a borrar el libro con el nombre: ' + nombre)
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
+    else:
+        return make_response(jsonify({'success': False}), 400)
 # endregion
 
-#DONE
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Edicion
 @app.route('/insert_edicion', methods=['POST'])
 def insert_edicion():
@@ -131,18 +165,18 @@ def insert_edicion():
     codigo = data['code']
     año = data['año']
     lengua = data['language']
+
+    if codigo == '' or año == '' or lengua == '':
+        return make_response(jsonify({'success': False}), 500)
+
     ediciones = db['Edicion']
 
     if ediciones.find_one({"ISBN": codigo}):
         print('La edicion ya existe')
+        return make_response(jsonify({'success': False}), 400)
     else:
         ediciones.insert_one({"ISBN": codigo, "año": año, "Idioma": lengua})
-
-    #Aquí se insertaría el libro en la base de datos
-    #print('El codigo de la Edicion ingresado es: ' + codigo)
-    #print('El año de la Edicion ingresado es: ' + año)
-    #print('La lengua de la Edicion ingresado es: ' + lengua)
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
 
 @app.route('/get_edicion', methods=['POST'])
 def get_edicion():
@@ -155,8 +189,6 @@ def get_edicion():
         return jsonify({'field1': ans['ISBN'], 'field2': ans['año'], 'field3': ans['Idioma']})
     else:
         return jsonify({'field1': '', 'field2': '', 'field3': ''})
-    
-    print("Los resultados del query son: " + result1 + ", " + result2 + ", " + result3)
 
 @app.route('/update_edicion', methods=['POST'])
 def update_edicion():
@@ -165,6 +197,10 @@ def update_edicion():
     codigo = data['code']
     año = data['año']
     lengua = data['language']
+
+    if old == '' or codigo == '' or año == '' or lengua == '':
+        return make_response(jsonify({'success': False}), 500)
+
     ediciones = db['Edicion']
     libros = db['Libro']
 
@@ -174,29 +210,29 @@ def update_edicion():
         for libro in libros_cod:
             libros.update_one({"_id": libro["_id"]}, {"$set": {"ISBN": codigo}})
         ediciones.update_one({"ISBN": old}, {"$set": {"ISBN": codigo, "año": año, "Idioma": lengua}})
-
-    #Aquí se insertaría el libro en la base de datos
-    #print('El codigo de la Edicion ingresado es: ' + codigo)
-    #print('El año de la Edicion ingresado es: ' + año)
-    #print('La lengua de la Edicion ingresado es: ' + lengua)
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
+    else:
+        return make_response(jsonify({'success': False}), 400)
 
 @app.route('/delete_edicion', methods=['POST'])
 def delete_edicion():
     data = request.get_json()
     nombre = data['name']
+
+    if nombre == '':
+        return make_response(jsonify({'success': False}), 500)
+
     ediciones = db['Edicion']
 
     edicion = ediciones.find_one({"ISBN": nombre})
     if edicion:
         ediciones.delete_one({"ISBN": nombre})
-
-    #Aquí se borraría el autor de la base de datos
-    print('Se va a borrar la edicion con el codigo: ' + nombre)
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
+    else:
+        return make_response(jsonify({'success': False}), 400)
 # endregion
 
-#DONE   
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Copia
 @app.route('/insert_copia', methods=['POST'])
 def insert_copia():
@@ -205,6 +241,10 @@ def insert_copia():
     copias = db['Copia']
     numero = data['numero']
     codigo = data['codigo']
+
+    if numero == '' or codigo == '':
+        return make_response(jsonify({'success': False}), 500)
+
     ediciones = db['Edicion']
     
     if not ediciones.find_one({'ISBN': codigo}):
@@ -213,18 +253,18 @@ def insert_copia():
     
     if copias.find_one({'numero': numero}):
         print('La copia ya existe')
-        response = make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
+        return make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
     else:
         copias.insert_one({'numero': numero, 'ISBN': codigo})
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
-    return response
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
 
 @app.route('/delete_copia', methods=['POST'])
 def delete_copia():
     data = request.get_json()
     num = data['num']
-    autores = db['Autor']
-    libros = db['Libro']
+
+    if num == '':
+        return make_response(jsonify({'success': False}), 500)
 
     copias = db['Copia']
 
@@ -234,26 +274,26 @@ def delete_copia():
         return make_response(jsonify({'success': True}), 200)
     else:
         return make_response(jsonify({'success': False}), 401)
-   
-    #Aquí se borraría el autor de la base de datos
-    #print('Se va a borrar el autor con el nombre: ' + nombre)
 # endregion
 
-#DONE
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Usuario
 @app.route('/insert_user', methods=['POST'])
 def insert_user():
     data = request.get_json()
     nombre = data['name']
+
+    if nombre == '':
+        return make_response(jsonify({'success': False}), 500)
+    
     rut = data['id']
 
     usuarios = db['Usuario']
     if usuarios.find_one({"rut": rut}):
-        response = make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
+        return make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
     else:
         usuarios.insert_one({"rut": rut, "name": nombre})
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
-    return response
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
 
 @app.route('/get_user', methods=['POST'])
 def get_user():
@@ -267,10 +307,6 @@ def get_user():
         return jsonify({'field1': ans['rut'], 'field2': ans['name']})
     else:
         return jsonify({'field1': '', 'field2': ''})
-    
-    result1, result2, result3 = 'Texto 1', 'Texto 2', 'Texto 3'
-    print("Los resultados del query son: " + result1 + ", " + result2 + ", " + result3)
-    return jsonify({'field1': result1, 'field2': result2, 'field3': result3})
 
 @app.route('/update_user', methods=['POST'])
 def update_usuario():
@@ -282,18 +318,31 @@ def update_usuario():
     usuario = db['Usuario']
     prestamo = db['Prestamo']
 
+    if prevRut == '' or rut == '' or name == '':
+        return make_response(jsonify({'success': False}), 500)
+
+    if usuario.find_one({"rut": rut}):  # Si ya existe un usuario con el nuevo RUT
+        return make_response(jsonify({'success': False}), 402)
+
+    
+
     us = usuario.find_one({"rut": prevRut})
     if us:
         prestamos_usuario = prestamo.find({"rut": prevRut})
         for prest in prestamos_usuario:
             prestamo.update_one({"_id": prest["_id"]}, {"$set": {"rut": rut}})
         usuario.update_one({"rut": prevRut}, {"$set": {"rut": rut, "name": name}})
-    return jsonify({'success': True})
+        return make_response(jsonify({'success': True}), 200)
+    else:
+        return make_response(jsonify({'success': False}), 400)
 
 @app.route('/delete_user', methods=['POST'])
 def delete_user():
     data = request.get_json()
     rut = data['rut']
+
+    if rut == '':
+        return make_response(jsonify({'success': False}), 500)
 
     usuario = db['Usuario']
     prestamo = db['Prestamo']
@@ -306,15 +355,13 @@ def delete_user():
         for prest in prestamos_usuario:
             prestamo.delete_one({"rut": rut})
         usuario.delete_one({"rut": rut})
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
     else:
         print('No se encontró ningún usuario con el ID proporcionado')
-        response = make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
-
-    return response
+        return make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
 # endregion
 
-#DONE
+#DONE!!!!
 # region Insertar/Actualizar/Borrar Prestamo
 @app.route('/insert_prestamo', methods=['POST'])
 def insert_prestamo():
@@ -324,6 +371,9 @@ def insert_prestamo():
     fechaFin = data['fechaDev']
     rut = data['rut']
     numCopia = data['copia']
+
+    if fechaInicio == '' or fechaFin == '' or rut == '' or numCopia == '':
+        return make_response(jsonify({'success': False}), 500)
 
     prestamos = db['Prestamo']
     usuarios = db['Usuario']
@@ -339,15 +389,15 @@ def insert_prestamo():
 
     if usuario and copia and disponible:
         prestamos.insert_one({"fechaPrestamo": fechaInicio, "fechaDev": fechaFin, "rut": rut, "copia": numCopia})
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
     else:
         if not usuario:
-            response = make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
+            return make_response(jsonify({'success': False}), 400)  # 400 is for Bad Request
         if not copia:
-            response = make_response(jsonify({'success': False}), 401)  # 401 is for Unauthorized
+            return make_response(jsonify({'success': False}), 401)  # 401 is for Unauthorized
         if not disponible:
-            response = make_response(jsonify({'success': False}), 402) # 402 is for Payment Required
-    return response
+            return make_response(jsonify({'success': False}), 402) # 402 is for Payment Required
+    return make_response(jsonify({'success': True}), 200)  # 200 is for OK
 
 @app.route('/get_prestamo', methods=['POST'])
 def get_prestamo():
@@ -370,25 +420,28 @@ def update_prestamo():
     fecha1 = data['date1']
     fecha2 = data['date2']
 
+    if numCopia == '' or fecha1 == '' or fecha2 == '':
+        return make_response(jsonify({'success': False}), 500)
 
     prestamos = db['Prestamo']
-    copias = db['Copia']
 
     prestamo = prestamos.find_one({"copia": numCopia})
 
     if prestamo:
         prestamos.update_one({"copia": numCopia}, {"$set": {"fechaPrestamo": fecha1, "fechaDev": fecha2}})
         print('Prestamo actualizado con éxito')
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
     else:
         print('Prestamo no encontrado')
-        response = make_response(jsonify({'success': False}), 400)
-    return response
+        return make_response(jsonify({'success': False}), 400)
 
 @app.route('/delete_prestamo', methods=['POST'])
 def delete_prestamo():
     data = request.get_json()
     id = data['id']
+
+    if id == '':
+        return make_response(jsonify({'success': False}), 500)
 
     prestamos = db['Prestamo']
     copias = db['Copia']
@@ -397,10 +450,9 @@ def delete_prestamo():
 
     if prestamo:
         prestamos.delete_one({"copia": id})
-        response = make_response(jsonify({'success': True}), 200)  # 200 is for OK
+        return make_response(jsonify({'success': True}), 200)  # 200 is for OK
     else:
-        response = make_response(jsonify({'success': False}), 400)
-    return response
+        return make_response(jsonify({'success': False}), 400)
 # endregion
 
 if __name__ == '__main__':
